+++ scripts/__init__.py
@@ -1,0 +1,1 @@
+"""M99 operational scripts package."""